split out into more modules, remove executor, fix query caching, cleanup
