+++ src/elem.rs
@@ -1,19 +1,14 @@
 use crate::arbitrary::{ArbitraryNumber, ArbitraryMap, ArbitraryValue};
-use crate::stack::{Location};
-
-use thiserror::Error;
 
 use std::cmp;
 use std::fmt;
 use std::fmt::{Debug, Display, Formatter};
-use std::marker::PhantomData;
-use std::iter::{FromIterator, IntoIterator};
-
-use serde::{Deserialize, Serialize};
-use serde_json::{Map, Number, Value};
+use std::iter::IntoIterator;
 
 use enumset::{EnumSet, EnumSetType};
 use quickcheck::{empty_shrinker, Arbitrary, Gen};
+use serde::{Deserialize, Serialize};
+use serde_json::{Map, Number, Value};
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub enum Elem {
@@ -127,6 +122,7 @@ impl Arbitrary for Elem {
         symbol.arbitrary_contents(g)
     }
 
+    // TODO: shrink
     fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
         empty_shrinker()
         // let self_copy = self.clone();
@@ -218,467 +214,5 @@ impl Elem {
     pub fn symbol_str(&self) -> &'static str {
       From::from(self.symbol())
     }
-}
-
-
-
-
-
-
-/* #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)] */
-/* pub enum BaseElemType { */
-/*     Any, */
-/*     Concat, */
-/*     Index, */
-/*     Slice, */
-/*     ElemSymbol(ElemSymbol), */
-/* } */
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct ElemTypeInfo {
-    /* base_elem_type: BaseElemType, */
-    location: Location,
-}
-
-// TODO: make fields private?
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct ElemType {
-    pub type_set: EnumSet<ElemSymbol>,
-    pub info: Vec<ElemTypeInfo>,
-}
-
-// Formatting:
-// ```
-// ElemType {
-//     type_set: {A, B, C},
-//     info: _,
-// }
-// ```
-//
-// Results in:
-// ```
-// {A, B, C}
-// ```
-impl Display for ElemType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f,
-               "{{{}}}",
-               self.type_set.iter()
-               .fold(String::new(),
-                     |memo, x| {
-                         let x_str: &'static str = From::from(x);
-                         if memo == "" {
-                            x_str.to_string()
-                         } else {
-                            memo + ", " + &x_str.to_string()
-                         }
-                    }
-               ))
-    }
-}
-
-#[cfg(test)]
-mod elem_type_display_tests {
-    use super::*;
-
-    #[test]
-    fn test_empty() {
-        let elem_type = ElemType {
-            type_set: EnumSet::empty(),
-            info: vec![],
-        };
-        assert_eq!("{}", format!("{}", elem_type));
-    }
-
-    #[test]
-    fn test_singleton() {
-        for elem_symbol in EnumSet::all().iter() {
-            let elem_type = ElemType {
-                type_set: EnumSet::only(elem_symbol),
-                info: vec![],
-            };
-            assert_eq!(format!("{{{}}}", Into::<&'static str>::into(elem_symbol)),
-                       format!("{}", elem_type));
-        }
-    }
-
-    #[test]
-    fn test_all() {
-        assert_eq!("{Unit, Bool, Number, Bytes, String, Array, Object, JSON}",
-                   format!("{}", ElemType::any(vec![])));
-    }
-}
-
-impl ElemSymbol {
-    pub fn elem_type(&self, locations: Vec<Location>) -> ElemType {
-        ElemType {
-            type_set: EnumSet::only(*self),
-            info: locations.iter()
-                .map(|&location|
-                     ElemTypeInfo {
-                         // base_elem_type: BaseElemType::ElemSymbol(*self),
-                         location: location,
-                    }).collect(),
-        }
-    }
-}
-
-impl Elem {
-    pub fn elem_type(&self, locations: Vec<Location>) -> ElemType {
-        self.symbol().elem_type(locations)
-    }
-}
-
-impl ElemType {
-    pub fn from_locations(type_set: EnumSet<ElemSymbol>,
-                          locations: Vec<Location>) -> Self {
-        ElemType {
-            type_set: type_set,
-            info: locations.iter()
-                .map(|&location|
-                     ElemTypeInfo {
-                         location: location,
-                    }).collect(),
-        }
-    }
-
-    pub fn any(locations: Vec<Location>) -> Self {
-        Self::from_locations(
-            EnumSet::all(),
-            locations)
-    }
-
-    pub fn union(&self, other: Self) -> Result<Self, ElemTypeError> {
-        let both = self.type_set.union(other.type_set);
-        let mut both_info = self.info.clone();
-        both_info.append(&mut other.info.clone());
-        Ok(ElemType {
-            type_set: both,
-            info: both_info,
-        })
-    }
-
-    pub fn unify(&self, other: Self) -> Result<Self, ElemTypeError> {
-        let both = self.type_set.intersection(other.type_set);
-        if both.is_empty() {
-            Err(ElemTypeError::UnifyEmpty {
-                lhs: self.clone(),
-                rhs: other.clone(),
-            })
-        } else {
-            let mut both_info = self.info.clone();
-            both_info.append(&mut other.info.clone());
-            Ok(ElemType {
-                type_set: both,
-                info: both_info,
-            })
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Error)]
-pub enum ElemTypeError {
-    #[error("ElemType::unify applied to non-intersecting types:\nlhs:\n{lhs}\nrhs:\n{rhs}")]
-    UnifyEmpty {
-        lhs: ElemType,
-        rhs: ElemType,
-    },
-}
-
-
-
-// BEGIN DebugAsDisplay
-#[derive(Clone, PartialEq, Eq)]
-struct DebugAsDisplay<T>
-where
-    T: Display,
-{
-    t: T,
-}
-
-impl<T> Display for DebugAsDisplay<T>
-where
-    T: Display,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f, "{}", self.t)
-    }
-}
-
-impl<T> Debug for DebugAsDisplay<T>
-where
-    T: Display,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f, "{}", self.t)
-    }
-}
-// END DebugAsDisplay
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct StackType {
-    pub types: Vec<ElemType>,
-}
-
-impl IntoIterator for StackType {
-    type Item = ElemType;
-    type IntoIter = <Vec<ElemType> as IntoIterator>::IntoIter;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.types.into_iter()
-    }
-}
-
-impl FromIterator<ElemType> for StackType {
-    fn from_iter<T>(iter: T) -> Self
-    where
-        T: IntoIterator<Item = ElemType>,
-    {
-        StackType {
-            types: FromIterator::from_iter(iter),
-        }
-    }
-}
-
-impl StackType {
-    pub fn len(&self) -> usize {
-        self.types.len()
-    }
-
-    pub fn push(&mut self, elem_type: ElemType) -> () {
-        self.types.insert(0, elem_type)
-    }
-
-    pub fn push_n(&mut self, elem_type: ElemType, count: usize) -> () {
-        for _index in 0..count {
-            self.push(elem_type.clone())
-        }
-    }
-}
-
-// Uses DebugAsDisplay to eliminate '"' around strings:
-// ["{Number}", "{Array, Object}"] -> [{Number}, {Array, Object}]
-impl Display for StackType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        f.debug_list()
-            .entries(self.types
-                     .iter()
-                     .map(|x| DebugAsDisplay { t: format!("{}", x) }))
-            .finish()?;
-        Ok(())
-    }
-}
-
-
-
-
-
-
-pub trait AnElem: Clone + std::fmt::Debug + PartialEq {
-    // TODO: rename
-
-    // fn elem_symbol(t: PhantomData<Self>) -> ElemType;
-    fn elem_symbol(t: PhantomData<Self>) -> EnumSet<ElemSymbol>;
-    fn to_elem(self) -> Elem;
-    fn from_elem(t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError>;
-}
-
-impl AnElem for Elem {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::all()
-    }
-
-    fn to_elem(self) -> Elem {
-        self
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        Ok(x)
-    }
-}
-
-
-impl AnElem for () {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Unit)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Unit
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Unit => Ok(()),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for bool {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Bool)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Bool(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Bool(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for Number {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Number)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Number(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Number(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for Vec<u8> {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Bytes)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Bytes(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Bytes(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for String {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::String)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::String(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::String(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for Vec<Value> {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Array)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Array(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Array(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for Map<String, Value> {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Object)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Object(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Object(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-impl AnElem for Value {
-    fn elem_symbol(_t: PhantomData<Self>) -> EnumSet<ElemSymbol> {
-        EnumSet::only(ElemSymbol::Json)
-    }
-
-    fn to_elem(self) -> Elem {
-        Elem::Json(self)
-    }
-
-    fn from_elem(_t: PhantomData<Self>, x: Elem) -> Result<Self, AnElemError> {
-        let elem_symbol = <Self as AnElem>::elem_symbol(PhantomData);
-        match x {
-            Elem::Json(y) => Ok(y),
-            _ => Err(AnElemError::UnexpectedElemType {
-                expected: elem_symbol,
-                found: x,
-            }),
-        }
-    }
-}
-
-
-#[derive(Clone, Debug, Error)]
-pub enum AnElemError {
-    #[error("AnElem::from_elem: element popped from the stack\n\n{found}\n\nwasn't the expected type:\n{expected:?}")]
-    UnexpectedElemType {
-        expected: EnumSet<ElemSymbol>,
-        found: Elem,
-    },
-
-    #[error("<Or<_, _> as AnElem>::from_elem: {e_hd:?}\n{e_tl:?}")]
-    PopOr {
-        e_hd: Box<Self>,
-        e_tl: Box<Self>,
-    },
 }
