+++ src/lib.rs
@@ -4,21 +4,34 @@ mod arbitrary;
 pub use arbitrary::{ArbitraryNumber, ArbitraryMap, ArbitraryValue};
 mod elem;
 pub use elem::{Elem, ElemSymbol};
+mod elem_type;
+pub use elem_type::{ElemType, StackType};
+mod an_elem;
+pub use an_elem::{AnElem, AnElemError};
+mod location;
+pub use location::{LineNo};
 mod stack;
 pub use stack::{Stack, StackError};
 mod types;
 mod types_scratch;
-pub use types_scratch::{Instruction, Instrs, AssertTrue, Push, Lookup, UnpackJson, Index, CheckEq, StringEq};
 mod json_template;
 pub use json_template::{TMap, TValue, TValueRunError, Template};
 mod query;
 pub use query::{Query, QueryType, QueryError};
-mod instruction;
-pub use instruction::{Instructions};
+mod untyped_instruction;
+pub use untyped_instruction::{Instruction};
+mod untyped_instructions;
+pub use untyped_instructions::{Instructions};
+mod typed_instruction;
+pub use typed_instruction::{IsInstructionT};
+mod typed_instructions;
+pub use typed_instructions::{AssertTrue, Push, Lookup, UnpackJson, Index, CheckEq, StringEq};
+mod typed_instr;
+pub use typed_instr::Instr;
+mod typed_instrs;
+pub use typed_instrs::Instrs;
 mod parse;
 pub use parse::{parse, parse_json};
-mod executor;
-pub use executor::Executor;
 
 mod rest_api;
 pub use rest_api::Api;
