+++ src/location.rs
@@ -1,0 +1,41 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+pub struct LineNo {
+    pub line_no: usize,
+}
+
+impl From<usize> for LineNo {
+    fn from(line_no: usize) -> Self {
+        LineNo {
+            line_no: line_no,
+        }
+    }
+}
+
+pub type ArgumentIndex = usize;
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+pub struct Location {
+    line_no: LineNo,
+    argument_index: ArgumentIndex,
+    is_input: bool,
+}
+
+impl LineNo {
+    pub fn in_at(&self, argument_index: usize) -> Location {
+        Location {
+            line_no: *self,
+            argument_index: argument_index,
+            is_input: true,
+        }
+    }
+
+    pub fn out_at(&self, argument_index: usize) -> Location {
+        Location {
+            line_no: *self,
+            argument_index: argument_index,
+            is_input: false,
+        }
+    }
+}
