+++ src/stack.rs
@@ -1,26 +1,18 @@
-// use crate::restack::{RestackError};
-use crate::elem::{Elem, StackType, AnElem, AnElemError, ElemSymbol};
-
-// use std::collections::BTreeMap;
-// use std::cmp;
-// use std::iter::{FromIterator};
+use crate::elem::{Elem, ElemSymbol};
+use crate::elem_type::{StackType};
+use crate::an_elem::{AnElem, AnElemError};
+use crate::location::{LineNo};
 
 use std::fmt;
 use std::fmt::{Display, Formatter};
-// // use std::alloc::string;
-// use std::marker::PhantomData;
-// use std::sync::Arc;
 use std::marker::PhantomData;
 
 use enumset::{EnumSet};
 use serde::{Deserialize, Serialize};
-// use serde_json::{Map, Number, Value};
 use thiserror::Error;
 use generic_array::{GenericArray, ArrayLength};
 use typenum::marker_traits::Unsigned;
 
-
-// TODO: use for execution
 // TODO: pub field needed?
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Serialize, Deserialize)]
 pub struct Stack {
@@ -61,7 +53,6 @@ impl Stack {
 
     pub fn push(&mut self, elem: Elem) {
         let mut memo = vec![elem];
-        // memo.append(&mut self.stack.clone());
         memo.append(&mut self.stack);
         self.stack = memo;
     }
@@ -104,7 +95,6 @@ impl Stack {
 }
 
 
-
 #[derive(Clone, Debug, Error)]
 pub enum StackError {
     #[error("Stack::pop: tried to pop from an empty stack")]
@@ -134,46 +124,5 @@ pub enum StackError {
 impl From<AnElemError> for StackError {
     fn from(x: AnElemError) -> Self {
         Self::AnElemError(x)
-    }
-}
-
-// TODO: relocate LineNo, ArgumentIndex, Location
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct LineNo {
-    pub line_no: usize,
-}
-
-impl From<usize> for LineNo {
-    fn from(line_no: usize) -> Self {
-        LineNo {
-            line_no: line_no,
-        }
-    }
-}
-
-pub type ArgumentIndex = usize;
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct Location {
-    line_no: LineNo,
-    argument_index: ArgumentIndex,
-    is_input: bool,
-}
-
-impl LineNo {
-    pub fn in_at(&self, argument_index: usize) -> Location {
-        Location {
-            line_no: *self,
-            argument_index: argument_index,
-            is_input: true,
-        }
-    }
-
-    pub fn out_at(&self, argument_index: usize) -> Location {
-        Location {
-            line_no: *self,
-            argument_index: argument_index,
-            is_input: false,
-        }
     }
 }
