+++ src/types_scratch.rs
@@ -1,48 +1,35 @@
-use crate::elem::{Elem, ElemType, ElemTypeError, ElemSymbol, StackType, AnElem};
+use crate::restack::{RestackError};
+use crate::elem::{Elem, ElemSymbol};
+use crate::elem_type::{ElemType, ElemTypeError, StackType};
+use crate::an_elem::{AnElem};
 use crate::stack::{Stack, StackError};
-use crate::restack::{Restack, RestackError};
-use crate::types::{Context, ContextError, Type, TypeError, Empty, AnError, Nil};
+use crate::types::{Context, ContextError, Type, Nil};
 
-use std::cmp;
-use std::convert::TryFrom;
-// use std::iter::FromIterator;
 use std::marker::PhantomData;
-// use std::fmt;
 use std::fmt::Debug;
 use std::sync::{Arc, Mutex};
-use std::string::FromUtf8Error;
 
 use enumset::EnumSet;
 use generic_array::functional::FunctionalSequence;
 use generic_array::sequence::GenericSequence;
-use generic_array::typenum::{U0, U1, U2};
+use generic_array::typenum::{U0};
 use generic_array::{GenericArray, GenericArrayIter, ArrayLength};
-use serde::{Deserialize, Serialize};
 use serde_json::{Map, Number, Value};
 use thiserror::Error;
 use typenum::marker_traits::Unsigned;
 
-// use generic_array::typenum::{B1};
-// use typenum::marker_traits::Unsigned;
-// use typenum::type_operators::IsLess;
-
-// NEXT:
-// - delete old IsInstruction
-//
-// - add typing info as with pop-stack
-//  + get typing up to parity
-//  + add special-case unifier for restack + IsInstructionT for testing?
-//
+// TODO:
 // - random type -> ~random inhabitant of the type
-// - random typed program!?
+// - random typed program?
 
+// TODO: rename
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Singleton<T, N>
 where
     T: AnElem,
     N: ArrayLength<T> + Debug,
 {
-    array: GenericArray<T, N>,
+    pub array: GenericArray<T, N>,
 }
 
 impl<T, N> IntoIterator for Singleton<T, N>
@@ -95,6 +82,9 @@ pub enum ElemsPopError {
         error: Arc<Self>,
     },
 
+    #[error("Instr::run: ElemTypeError:\n{0}")]
+    RestackError(RestackError),
+
     #[error("Elems::elem_type (Or): Set includes repeated type:\n{0}")]
     ElemTypeError(ElemTypeError),
 
@@ -113,9 +103,6 @@ pub trait Elems: Clone + Debug + IntoIterator<Item = Elem> {
     // fn left(s: PhantomData<Self>, x: GenericArray<Self::Hd, Self::N>) -> Self;
     // fn right(s: PhantomData<Self>, x: Self::Tl) -> Self;
     fn or<T, F: Fn(&GenericArray<Self::Hd, Self::N>) -> T, G: Fn(&Self::Tl) -> T>(&self, f: F, g: G) -> T;
-
-    // fn to_elems(self) -> Elem;
-    // fn from_elems(t: PhantomData<Self>, x: &mut Stack) -> Result<Self, ElemsError>;
 
     fn pop(_x: PhantomData<Self>, stack: &mut Stack) -> Result<Self, ElemsPopError>
     where
@@ -236,7 +223,6 @@ where
     Right(U),
 }
 
-// #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum IterOr<T, N, U>
 where
     T: AnElem,
@@ -342,7 +328,7 @@ where
 {}
 
 // TODO: AnElem: &self -> AllElems<U1>
-type AllElems<N> =
+pub type AllElems<N> =
     Or<(), N,
     Or<bool, N,
     Or<Number, N,
@@ -352,7 +338,7 @@ type AllElems<N> =
     Or<Map<String, Value>, N,
     Singleton<Value, N>>>>>>>>;
 
-fn all_elems_untyped<N>(x: &AllElems<N>) -> GenericArray<Elem, N>
+pub fn all_elems_untyped<N>(x: &AllElems<N>) -> GenericArray<Elem, N>
 where
     N: Debug +
     ArrayLength<()> +
@@ -404,8 +390,8 @@ where
     T: AnElem,
     N: ArrayLength<T> + Debug,
 {
-    singleton: Singleton<T, N>,
-    returning: Return<T>,
+    pub singleton: Singleton<T, N>,
+    pub returning: Return<T>,
 }
 
 impl<T, N> IntoIterator for ReturnSingleton<T, N>
@@ -550,7 +536,6 @@ where
                     Or::Right(y) => Self::Right(y),
                 }
             })
-
     }
 
     fn elem_type(_t: PhantomData<Self>) -> Result<ElemType, ElemsPopError> {
@@ -876,963 +861,8 @@ where
         // let num_elem_type_hd = <<T as Elems>::N as Unsigned>::to_usize();
         let mut type_hd = IOElems::type_of(PhantomData::<T>)?;
         let elem_type_tl = IsList::stack_type(PhantomData::<U>)?;
-
         type_hd.append_inputs(elem_type_tl);
         Ok(type_hd)
-    }
-}
-
-
-
-
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Serialize, Deserialize)]
-pub enum Instruction {
-    Push(Elem),
-    Restack(Restack),
-    HashSha256,
-    CheckLe,
-    CheckLt,
-    CheckEq,
-    StringEq,
-    BytesEq,
-    Concat,
-    Slice,
-    Index,
-    Lookup,
-    AssertTrue,
-    ToJson,
-    UnpackJson(ElemSymbol),
-    StringToBytes,
-}
-
-pub trait IsInstructionT: Debug {
-    type IO: IOList;
-    type Error: AnError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError>;
-    fn name(x: PhantomData<Self>) -> String;
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error>;
-}
-
-#[derive(Clone, Debug, Error)]
-pub enum StackInstructionError {
-    #[error("StackInstructionError::ElemsPopError:\n{0}")]
-    ElemsPopError(ElemsPopError),
-
-    #[error("RawStackInstructionError:\n{0}")]
-    RawStackInstructionError(String),
-
-    #[error("MissingOutput:\n{instruction}\n\n{stack_input}")]
-    // TODO: more granular error typing
-    MissingOutput {
-        instruction: String,
-        stack_input: String,
-    },
-
-    #[error("Instrs::type_of_mono type error:\n{0}")]
-    TypeError(TypeError),
-
-    #[error("StackInstructionError::RestackError:\n{0}")]
-    RestackError(RestackError),
-
-    #[error("StackInstructionError::DebugJsonError:\n{0}")]
-    DebugJsonError(Arc<serde_json::Error>),
-
-    #[error("UnpackJsonNotSingleton:\n{first_value:?}\n{second_value:?}")]
-    UnpackJsonNotSingleton {
-        first_value: Option<ElemSymbol>,
-        second_value: Option<ElemSymbol>,
-    },
-
-}
-
-pub trait IsStackInstruction: Debug {
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError>;
-    fn name(&self) -> String;
-    fn type_of(&self) -> Result<Type, ElemsPopError>;
-    fn stack_run(&self, stack: &mut Stack) -> Result<(), StackInstructionError>;
-}
-
-impl<T> IsStackInstruction for T
-where
-    T: IsInstructionT,
-{
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        self.to_instruction()
-    }
-
-    fn name(&self) -> String {
-        IsInstructionT::name(PhantomData::<Self>)
-    }
-
-    fn type_of(&self) -> Result<Type, ElemsPopError> {
-        IOList::type_of(PhantomData::<<T as IsInstructionT>::IO>)
-    }
-
-    fn stack_run(&self, stack: &mut Stack) -> Result<(), StackInstructionError> {
-        let stack_input = &IsList::pop(PhantomData::<<T as IsInstructionT>::IO>, stack)
-            .map_err(|e| StackInstructionError::ElemsPopError(e))?;
-        self.run(stack_input)
-            .map_err(|e| StackInstructionError::RawStackInstructionError(format!("{:?}", e)))?;
-        let output_value = stack_input
-            .returning()
-            .ok_or_else(|| StackInstructionError::MissingOutput {
-                instruction: format!("{:?}", self),
-                stack_input: format!("{:?}", stack_input),
-            })?;
-        stack.push(output_value);
-        Ok(())
-    }
-}
-
-
-
-#[derive(Clone, Debug)]
-pub enum Instr {
-    Instr(Arc<dyn IsStackInstruction>),
-    Restack(Restack),
-}
-
-impl Instr {
-    pub fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        match self {
-            Self::Instr(instr) => instr.to_instruction(),
-            Self::Restack(restack) => Ok(Instruction::Restack(restack.clone())),
-        }
-    }
-}
-
-
-#[derive(Clone, Debug)]
-pub struct Instrs {
-    pub instrs: Vec<Instr>,
-}
-
-impl Instrs {
-    pub fn new() -> Self {
-        Instrs {
-            instrs: vec![],
-        }
-    }
-
-    /// Assuming an input stack of [Json, Json, ..] (num_input_json count),
-    /// what's the monomorphic type of Self?
-    pub fn type_of_mono(&self, num_input_json: usize) -> Result<StackType, StackInstructionError> {
-        let mut stack_type = (0..num_input_json).map(|_| ElemType::from_locations(EnumSet::only(ElemSymbol::Json), vec![])).collect();
-        for (line_no, instr_or_restack) in (&self.instrs).into_iter().enumerate() {
-            println!("------------------------------------------------------------------------------------------");
-            println!("line_no: {}", line_no);
-            println!("{:?}\n", instr_or_restack);
-            match instr_or_restack {
-                Instr::Instr(instr) => {
-                    let mut instr_type = instr.type_of()
-                        .map_err(|e| StackInstructionError::ElemsPopError(e))?;
-                    println!("instr: {}\n", instr_type);
-                    stack_type = instr_type.specialize_to_input_stack(stack_type)
-                        .map_err(|e| StackInstructionError::TypeError(e))?;
-                },
-                Instr::Restack(restack) => {
-                    restack.run(&mut stack_type.types)
-                        .map_err(|e| StackInstructionError::RestackError(e))?
-                },
-            }
-        }
-        println!("------------------------------------------------------------------------------------------");
-        println!("Finished running successfully.\n");
-        println!("Final stack:");
-        Ok(stack_type)
-    }
-
-    pub fn run(&self, stack: &mut Stack) -> Result<(), StackInstructionError> {
-        for (line_no, instr_or_restack) in (&self.instrs).into_iter().enumerate() {
-            stack.debug().map_err(|e| StackInstructionError::DebugJsonError(Arc::new(e)))?;
-            println!("------------------------------------------------------------------------------------------");
-            println!("line_no: {}", line_no);
-            println!("{:?}\n", instr_or_restack);
-            match instr_or_restack {
-                Instr::Instr(instr) => {
-                    println!("");
-                    stack.debug_type();
-                    match instr.type_of() {
-                        Ok(instr_type) => {
-                            println!("instr: {}\n", instr_type);
-                            let mut mut_instr_type = instr_type.clone();
-                            match mut_instr_type
-                                .specialize_to_input_stack(stack.type_of()) {
-                                                           // .clone()
-                                                           // .stack
-                                                           // .into_iter()
-                                                           // .map(|x| x.elem_type(vec![]))
-                                                           // .collect()) {
-                                Ok(_) => println!("specialized: {}\n", mut_instr_type),
-                                Err(e) => println!("specialization failed:\n{}\n", e),
-                            }
-                        },
-                        Err(e) => println!("instr type_of errror: {}\n", e),
-                    }
-                    println!("");
-                    instr.stack_run(stack)?
-                },
-                Instr::Restack(restack) => {
-                    restack.run(&mut stack.stack)
-                        .map_err(|e| StackInstructionError::RestackError(e))?
-                },
-            }
-        }
-        println!("------------------------------------------------------------------------------------------");
-        println!("Finished running successfully.\n");
-        println!("Final stack:");
-        stack.debug().map_err(|e| StackInstructionError::DebugJsonError(Arc::new(e)))?;
-        Ok(())
-    }
-
-    pub fn instr(&mut self, instr: impl IsStackInstruction + 'static) -> () {
-        self.instrs.push(Instr::Instr(Arc::new(instr)))
-    }
-
-    pub fn restack(&mut self, restack: Restack) -> () {
-        self.instrs.push(Instr::Restack(restack))
-    }
-}
-
-
-
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Concat {}
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct ConcatError {}
-impl AnError for ConcatError {}
-
-// TODO: add string!
-// (Self::String(x), Self::String(y)) => {
-//     Ok(Self::String(String::from_utf8(Self::concat_generic(Vec::from(x.clone()), Vec::from(y.clone())))
-//                     .map_err(|_| ElemError::ConcatInvalidUTF8 { lhs: x, rhs: y })?))
-// },
-//
-// bytes, array, object
-impl IsInstructionT for Concat {
-    type IO = ConsOut<ReturnOr<Vec<u8>,             U2,
-                      ReturnOr<Vec<Value>,          U2,
-               ReturnSingleton<Map<String, Value>,  U2>>>, Nil>;
-    type Error = ConcatError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::Concat)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "concat".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let y = x.clone().hd();
-        match y {
-            ReturnOr::Left { array, returning } => {
-                let lhs = &array[0];
-                let rhs = &array[1];
-                returning.returning(lhs.into_iter().chain(rhs.into_iter()).cloned().collect());
-            },
-            ReturnOr::Right(ReturnOr::Left { array, returning }) => {
-                let lhs = &array[0];
-                let rhs = &array[1];
-                returning.returning(lhs.into_iter().chain(rhs.into_iter()).cloned().collect());
-            },
-            ReturnOr::Right(ReturnOr::Right(ReturnSingleton { singleton, returning })) => {
-                let lhs = &singleton.array[0];
-                let rhs = &singleton.array[1];
-                returning.returning(lhs.into_iter().chain(rhs.into_iter()).map(|xy| (xy.0.clone(), xy.1.clone())).collect());
-            },
-        }
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct AssertTrue {}
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct AssertTrueError {}
-impl AnError for AssertTrueError {}
-
-impl IsInstructionT for AssertTrue {
-    type IO = ConsOut<ReturnSingleton<bool, U1>, Nil>;
-    // TODO: replace w/ Empty
-    type Error = AssertTrueError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::AssertTrue)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "assert_true".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let array = x.clone().hd().singleton.array;
-        let returning = x.clone().hd().returning;
-        if array[0] {
-            returning.returning(true);
-            Ok(())
-        } else {
-            Err(AssertTrueError {})
-        }
-    }
-}
-
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct Push<T: AnElem> {
-    pub push: T,
-}
-
-impl<T: AnElem> IsInstructionT for Push<T> {
-    type IO = ConsOut<ReturnSingleton<T, U0>, Nil>;
-    type Error = Empty;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::Push(self.push.clone().to_elem()))
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        format!("push_{:?}", AnElem::elem_symbol(PhantomData::<T>))
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        x.clone().hd().returning.returning(self.push.clone());
-        Ok(())
-    }
-}
-
-
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct HashSha256 {}
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct HashSha256Error {}
-impl AnError for HashSha256Error {}
-
-impl IsInstructionT for HashSha256 {
-    type IO = ConsOut<ReturnSingleton<Vec<u8>, U1>, Nil>;
-    type Error = Empty;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::HashSha256)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "sha256".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let array = x.clone().hd().singleton.array;
-        let returning = x.clone().hd().returning;
-        returning.returning(super::sha256(&array[0]));
-        Ok(())
-    }
-}
-
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Slice {}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum SliceError {
-    OffsetNotU64(Number),
-
-    LengthNotU64(Number),
-
-    Overflow {
-        offset: Number,
-        length: Number,
-    },
-
-    TooShort {
-        offset: usize,
-        length: usize,
-        iterable: String,
-    },
-
-    FromUtf8Error(FromUtf8Error),
-}
-
-impl From<FromUtf8Error> for SliceError {
-    fn from(error: FromUtf8Error) -> Self {
-        Self::FromUtf8Error(error)
-    }
-}
-
-impl AnError for SliceError {}
-
-// bytes, string, array, object
-impl IsInstructionT for Slice {
-    type IO = ConsOut<ReturnOr<Vec<u8>,             U1,
-                      ReturnOr<String,              U1,
-                      ReturnOr<Vec<Value>,          U1,
-               ReturnSingleton<Map<String, Value>,  U1>>>>,
-                Cons<Singleton<Number,              U2>, Nil>>;
-    type Error = SliceError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::Slice)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "slice".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let y = x.clone().hd();
-        let offset_length = x.clone().tl().hd().array;
-        let offset = &offset_length[0];
-        let length = &offset_length[1];
-        let u_offset = offset.as_u64()
-            .ok_or_else(|| SliceError::OffsetNotU64(offset.clone()))
-            .and_then(|x| usize::try_from(x).map_err(|_| SliceError::Overflow { offset: offset.clone(), length: length.clone() }))?;
-        let u_length = length.as_u64()
-            .ok_or_else(|| SliceError::LengthNotU64(length.clone()))
-            .and_then(|x| usize::try_from(x).map_err(|_| SliceError::Overflow { offset: offset.clone(), length: length.clone() }))?;
-        let u_offset_plus_length = u_offset.checked_add(u_length)
-            .ok_or_else(|| SliceError::Overflow { offset: offset.clone(), length: length.clone() })?;
-        match y.clone() {
-            ReturnOr::Left { array, returning } => {
-                let iterable = &array[0];
-                if iterable.clone().into_iter().count() < u_offset_plus_length {
-                    Err(())
-                } else {
-                    returning.returning(iterable.into_iter().skip(u_offset).take(u_length).copied().collect());
-                    Ok(())
-                }
-            },
-            ReturnOr::Right(ReturnOr::Left { array, returning }) => {
-                let iterable = &array[0];
-                if iterable.len() < u_offset_plus_length {
-                    Err(())
-                } else {
-                    returning.returning(String::from_utf8(Vec::from(iterable.clone()).into_iter().skip(u_offset).take(u_length).collect())?);
-                    Ok(())
-                }
-            },
-            ReturnOr::Right(ReturnOr::Right(ReturnOr::Left { array, returning })) => {
-                let iterable = &array[0];
-                if iterable.clone().into_iter().count() < u_offset_plus_length {
-                    Err(())
-                } else {
-                    returning.returning(iterable.into_iter().skip(u_offset).take(u_length).cloned().collect());
-                    Ok(())
-                }
-            },
-            ReturnOr::Right(ReturnOr::Right(ReturnOr::Right(ReturnSingleton { singleton: Singleton { array }, returning }))) => {
-                let iterable = &array[0];
-                if iterable.clone().into_iter().count() < u_offset_plus_length {
-                    Err(())
-                } else {
-                    returning.returning(iterable.into_iter().skip(u_offset).take(u_length).map(|xy| (xy.0.clone(), xy.1.clone())).collect());
-                    Ok(())
-                }
-            },
-        }.map_err(|_e| {
-            SliceError::TooShort {
-                offset: u_offset,
-                length: u_length,
-                // TODO: better error
-                iterable: format!("{:?}", y),
-            }
-        })
-    }
-}
-
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Index {}
-#[derive(Clone, Debug, PartialEq, Eq, Error)]
-pub enum IndexError {
-    #[error("Index: index not valid u64: {0:?}")]
-    IndexNotU64(Number),
-
-    #[error("Index: index not valid usize: {0:?}")]
-    Overflow(Number),
-
-    #[error("Index: iterable: {iterable:?}\nis too short for index: {index:?}")]
-    TooShort {
-        index: usize,
-        iterable: String,
-    },
-}
-impl AnError for IndexError {}
-
-// bytes, array, object
-impl IsInstructionT for Index {
-    type IO = ConsOut<ReturnSingleton<Value,                U0>,
-                       Cons<Singleton<Number,               U1>,
-                              Cons<Or<Vec<Value>,           U1,
-                            Singleton<Map<String, Value>,   U1>>, Nil>>>;
-    type Error = IndexError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::Index)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "index".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let index = x.clone().tl().hd().array[0].clone();
-        let y = &x.clone().tl().tl().hd();
-        let u_index = index.as_u64()
-            .ok_or_else(|| IndexError::IndexNotU64(index.clone()))
-            .and_then(|x| usize::try_from(x).map_err(|_| IndexError::Overflow(index.clone())))?;
-
-        let result = match y.clone() {
-            Or::Left(array) => {
-                array[0]
-                    .clone()
-                    .into_iter()
-                    .skip(u_index)
-                    .next()
-            },
-            Or::Right(Singleton { array }) => {
-                array[0]
-                    .clone()
-                    .into_iter()
-                    .skip(u_index)
-                    .next()
-                    .map(|(_x, y)| y)
-            },
-        }.ok_or_else(|| {
-            IndexError::TooShort {
-                index: u_index,
-                // TODO: better error
-                iterable: format!("{:?}", y),
-            }
-        })?;
-        returning.returning(result);
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct ToJson {}
-#[derive(Clone, Debug)]
-pub struct ToJsonError {
-    input: Elem,
-    error: Arc<serde_json::Error>,
-}
-impl AnError for ToJsonError {}
-
-impl IsInstructionT for ToJson {
-    type IO = ConsOut<ReturnSingleton<Value, U0>, Cons<AllElems<U1>, Nil>>;
-    type Error = ToJsonError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::ToJson)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "to_json".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let y = &x.clone().tl().hd();
-        let array = all_elems_untyped(y);
-        let z = array[0].clone();
-        returning.returning(serde_json::to_value(z.clone())
-                            .map_err(move |e| ToJsonError {
-                                input: z,
-                                error: Arc::new(e),
-        })?);
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Lookup {}
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct LookupError {
-    key: String,
-    map: Map<String, Value>,
-}
-impl AnError for LookupError {}
-
-impl IsInstructionT for Lookup {
-    type IO = ConsOut<ReturnSingleton<Value, U0>,
-                 Cons<Singleton<String, U1>,
-                 Cons<Singleton<Map<String, Value>, U1>, Nil>>>;
-    type Error = LookupError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::Lookup)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "lookup".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let key = &x.clone().tl().hd().array[0];
-        let map = &x.clone().tl().tl().hd().array[0];
-        returning.returning(map.get(key)
-           .ok_or_else(|| LookupError {
-               key: key.clone(),
-               map: map.clone(),
-           })?.clone());
-        Ok(())
-    }
-}
-
-
-#[derive(Debug)]
-pub struct UnpackJson<T: AnElem> {
-    pub t: PhantomData<T>,
-}
-#[derive(Debug)]
-pub struct UnpackJsonError {}
-impl AnError for UnpackJsonError {}
-
-pub trait AJsonElem: AnElem {
-    fn to_value(self) -> Value;
-    fn from_value(t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized;
-}
-
-impl AJsonElem for () {
-    fn to_value(self) -> Value {
-        Value::Null
-    }
-
-    fn from_value(_t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized {
-        match x {
-            Value::Null => Some(()),
-            _ => None,
-        }
-    }
-}
-
-impl AJsonElem for bool {
-    fn to_value(self) -> Value {
-        Value::Bool(self)
-    }
-
-    fn from_value(_t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized {
-        match x {
-            Value::Bool(y) => Some(y),
-            _ => None,
-        }
-    }
-}
-
-impl AJsonElem for Number {
-    fn to_value(self) -> Value {
-        Value::Number(self)
-    }
-
-    fn from_value(_t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized {
-        match x {
-            Value::Number(y) => Some(y),
-            _ => None,
-        }
-    }
-}
-
-impl AJsonElem for String {
-    fn to_value(self) -> Value {
-        Value::String(self)
-    }
-
-    fn from_value(_t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized {
-        match x {
-            Value::String(y) => Some(y),
-            _ => None,
-        }
-    }
-}
-
-impl AJsonElem for Vec<Value> {
-    fn to_value(self) -> Value {
-        Value::Array(self)
-    }
-
-    fn from_value(_t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized {
-        match x {
-            Value::Array(y) => Some(y),
-            _ => None,
-        }
-    }
-}
-
-impl AJsonElem for Map<String, Value> {
-    fn to_value(self) -> Value {
-        Value::Object(self)
-    }
-
-    fn from_value(_t: PhantomData<Self>, x: Value) -> Option<Self> where Self: Sized {
-        match x {
-            Value::Object(y) => Some(y),
-            _ => None,
-        }
-    }
-}
-
-impl<T: AJsonElem> IsInstructionT for UnpackJson<T> {
-    type IO = ConsOut<ReturnSingleton<T, U0>,
-                       Cons<Singleton<Value, U1>, Nil>>;
-    type Error = UnpackJsonError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        let mut symbol_set = <T as AnElem>::elem_symbol(PhantomData).into_iter();
-        match (symbol_set.next(), symbol_set.next()) {
-            (Some(elem_symbol), None) => Ok(Instruction::UnpackJson(elem_symbol)),
-            (x, y) => Err(StackInstructionError::UnpackJsonNotSingleton {
-                first_value: x,
-                second_value: y,
-            }),
-        }
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "unpack_json".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let json = &x.clone().tl().hd().array[0];
-        let result =
-            AJsonElem::from_value(PhantomData::<T>, json.clone())
-            .ok_or_else(|| UnpackJsonError {})?;
-        returning.returning(result);
-        Ok(())
-    }
-}
-
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct StringToBytes {}
-
-impl IsInstructionT for StringToBytes {
-    type IO = ConsOut<ReturnSingleton<Vec<u8>, U0>, Cons<Singleton<String, U1>, Nil>>;
-    type Error = Empty;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::StringToBytes)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "string_to_bytes".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let in_str = &x.clone().tl().hd().array[0];
-        returning.returning(in_str.clone().into_bytes());
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct CheckLe {}
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CheckLeError {
-    lhs: Elem,
-    rhs: Elem,
-}
-impl AnError for CheckLeError {}
-
-impl IsInstructionT for CheckLe {
-    type IO = ConsOut<ReturnSingleton<bool, U0>, Cons<AllElems<U2>, Nil>>;
-    type Error = CheckLeError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::CheckLe)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "check_le".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let y = &x.clone().tl().hd();
-        let array = all_elems_untyped(y);
-        let lhs = array[0].clone();
-        let rhs = array[1].clone();
-        let cmp_result = lhs.partial_cmp(&rhs)
-            .ok_or_else(|| CheckLeError {
-                lhs: lhs,
-                rhs: rhs
-        })?;
-        let result = match cmp_result {
-            cmp::Ordering::Less => true,
-            cmp::Ordering::Equal => true,
-            cmp::Ordering::Greater => false,
-        };
-        returning.returning(result);
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct CheckLt {}
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CheckLtError {
-    lhs: Elem,
-    rhs: Elem,
-}
-impl AnError for CheckLtError {}
-
-impl IsInstructionT for CheckLt {
-    type IO = ConsOut<ReturnSingleton<bool, U0>, Cons<AllElems<U2>, Nil>>;
-    type Error = CheckLtError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::CheckLt)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "check_lt".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let y = &x.clone().tl().hd();
-        let array = all_elems_untyped(y);
-        let lhs = array[0].clone();
-        let rhs = array[1].clone();
-        let cmp_result = lhs.partial_cmp(&rhs)
-            .ok_or_else(|| CheckLtError {
-                lhs: lhs,
-                rhs: rhs
-        })?;
-        let result = match cmp_result {
-            cmp::Ordering::Less => true,
-            _ => false,
-        };
-        returning.returning(result);
-        Ok(())
-    }
-}
-
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct CheckEq {}
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CheckEqError {
-    lhs: Elem,
-    rhs: Elem,
-}
-impl AnError for CheckEqError {}
-
-impl IsInstructionT for CheckEq {
-    type IO = ConsOut<ReturnSingleton<bool, U0>, Cons<AllElems<U2>, Nil>>;
-    type Error = CheckEqError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::CheckEq)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "check_eq".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let y = &x.clone().tl().hd();
-        let array = all_elems_untyped(y);
-        let lhs = array[0].clone();
-        let rhs = array[1].clone();
-        let cmp_result = lhs.partial_cmp(&rhs)
-            .ok_or_else(|| CheckEqError {
-                lhs: lhs,
-                rhs: rhs
-        })?;
-        let result = match cmp_result {
-            cmp::Ordering::Equal => true,
-            _ => false,
-        };
-        returning.returning(result);
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct StringEq {}
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct StringEqError {
-    lhs: String,
-    rhs: String,
-}
-impl AnError for StringEqError {}
-
-impl IsInstructionT for StringEq {
-    type IO = ConsOut<ReturnSingleton<bool, U0>, Cons<Singleton<String, U2>, Nil>>;
-    type Error = StringEqError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::StringEq)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "check_eq".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let array = &x.clone().tl().hd().array;
-        let lhs = array[0].clone();
-        let rhs = array[1].clone();
-        let cmp_result = lhs.partial_cmp(&rhs)
-            .ok_or_else(|| StringEqError {
-                lhs: lhs,
-                rhs: rhs
-        })?;
-        let result = match cmp_result {
-            cmp::Ordering::Equal => true,
-            _ => false,
-        };
-        returning.returning(result);
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct BytesEq {}
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct BytesEqError {
-    lhs: Vec<u8>,
-    rhs: Vec<u8>,
-}
-impl AnError for BytesEqError {}
-
-impl IsInstructionT for BytesEq {
-    type IO = ConsOut<ReturnSingleton<bool, U0>, Cons<Singleton<Vec<u8>, U2>, Nil>>;
-    type Error = BytesEqError;
-
-    fn to_instruction(&self) -> Result<Instruction, StackInstructionError> {
-        Ok(Instruction::BytesEq)
-    }
-
-    fn name(_x: PhantomData<Self>) -> String {
-        "check_eq".to_string()
-    }
-
-    fn run(&self, x: &Self::IO) -> Result<(), Self::Error> {
-        let returning = x.clone().hd().returning;
-        let array = &x.clone().tl().hd().array;
-        let lhs = array[0].clone();
-        let rhs = array[1].clone();
-        let cmp_result = lhs.partial_cmp(&rhs)
-            .ok_or_else(|| BytesEqError {
-                lhs: lhs,
-                rhs: rhs
-        })?;
-        let result = match cmp_result {
-            cmp::Ordering::Equal => true,
-            _ => false,
-        };
-        returning.returning(result);
-        Ok(())
     }
 }
