+++ src/untyped_instructions.rs
@@ -1,0 +1,54 @@
+use crate::untyped_instruction::{Instruction, InstructionError};
+use crate::typed_instruction::StackInstructionError;
+use crate::typed_instr::Instr;
+use crate::typed_instrs::Instrs;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Serialize, Deserialize)]
+pub struct Instructions {
+    pub instructions: Vec<Instruction>,
+}
+
+impl IntoIterator for Instructions {
+    type Item = Instruction;
+    type IntoIter = <Vec<Instruction> as std::iter::IntoIterator>::IntoIter;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.instructions.into_iter()
+    }
+}
+
+impl Instructions {
+    pub fn to_instrs(self) -> Result<Instrs, InstructionError> {
+        Ok(Instrs {
+            instrs: self.into_iter().map(|x| x.to_instr()).collect::<Result<Vec<Instr>, InstructionError>>()?,
+        })
+    }
+}
+
+impl Instrs {
+    pub fn to_instructions(self) -> Result<Instructions, StackInstructionError> {
+        Ok(Instructions {
+            instructions: self.instrs.into_iter().map(|x| x.to_instruction()).collect::<Result<Vec<Instruction>, StackInstructionError>>()?,
+        })
+    }
+}
+
+// Test program #1: [] -> []
+//
+// Instruction::Push(Elem::Bool(true)),
+// Instruction::Restack(Restack::id()),
+// Instruction::AssertTrue,
+
+// Test program #2
+//
+// ∀ (t0 ∊ {JSON}),
+// ∀ (t1 ∊ {JSON}),
+// ∀ (t2 ∊ {Object}),
+// [t1] ->
+// [t0, t2, t1]
+//
+// Instruction::Push(Elem::Json(Default::default())),
+// Instruction::UnpackJson(ElemSymbol::Object),
+// Instruction::Restack(Restack::dup()),
